# openapi.yaml
openapi: 3.1.0
info:
  title: AV Suite Backend Foundation
  description: Rehabilitation exercises management backend platform
  version: 0.1.0
paths:
...
  /api/v1/appointments:
    post:
      tags:
      - Appointments
      summary: Create Appointment
      description: Create a new staff-scheduled appointment for the authenticated
        clinic.
      operationId: create_appointment_api_v1_appointments_post
      security:
      - OAuth2PasswordBearer: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AppointmentCreate'
      responses:
        '201':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AppointmentResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
    get:
      tags:
      - Appointments
      summary: List Appointments
      description: List appointments for the authenticated clinic with filtering by
        date, therapist, or patient.
      operationId: list_appointments_api_v1_appointments_get
      security:
      - OAuth2PasswordBearer: []
      parameters:
      - name: offset
        in: query
        required: false
        schema:
          type: integer
          minimum: 0
          default: 0
          title: Offset
      - name: limit
        in: query
        required: false
        schema:
          type: integer
          maximum: 500
          minimum: 1
          default: 100
          title: Limit
      - name: date
        in: query
        required: false
        schema:
          anyOf:
          - type: string
            format: date
          - type: 'null'
          title: Date
      - name: therapist
        in: query
        required: false
        schema:
          anyOf:
          - type: string
            format: uuid
          - type: 'null'
          title: Therapist
      - name: patient
        in: query
        required: false
        schema:
          anyOf:
          - type: string
            format: uuid
          - type: 'null'
          title: Patient
      - name: status
        in: query
        required: false
        schema:
          anyOf:
          - $ref: '#/components/schemas/AppointmentStatus'
          - type: 'null'
          title: Status
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AppointmentListResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/v1/appointments/{id}:
    get:
      tags:
      - Appointments
      summary: Get Appointment
      description: Retrieve an appointment by ID for the authenticated clinic.
      operationId: get_appointment_api_v1_appointments__id__get
      security:
      - OAuth2PasswordBearer: []
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
          format: uuid
          title: Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AppointmentResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
    patch:
      tags:
      - Appointments
      summary: Update Appointment
      description: Reschedule or update appointment status/details for the authenticated
        clinic.
      operationId: update_appointment_api_v1_appointments__id__patch
      security:
      - OAuth2PasswordBearer: []
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
          format: uuid
          title: Id
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AppointmentUpdate'
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AppointmentResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
    delete:
      tags:
      - Appointments
      summary: Soft Cancel Appointment
      description: Soft-cancel an appointment for the authenticated clinic.
      operationId: soft_cancel_appointment_api_v1_appointments__id__delete
      security:
      - OAuth2PasswordBearer: []
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
          format: uuid
          title: Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AppointmentResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
...
components:
  schemas:
...
    AppointmentCreate:
      properties:
        patient_id:
          type: string
          format: uuid
          title: Patient Id
        therapist_id:
          type: string
          format: uuid
          title: Therapist Id
        scheduled_at:
          type: string
          format: date-time
          title: Scheduled At
        duration_minutes:
          type: integer
          exclusiveMinimum: 0.0
          title: Duration Minutes
          default: 30
        status:
          allOf:
          - $ref: '#/components/schemas/AppointmentStatus'
          default: scheduled
        source:
          allOf:
          - $ref: '#/components/schemas/AppointmentSource'
          default: manual
      type: object
      required:
      - patient_id
      - therapist_id
      - scheduled_at
      title: AppointmentCreate
      description: Payload for creating a new appointment.
    AppointmentListResponse:
      properties:
        items:
          items:
            $ref: '#/components/schemas/AppointmentResponse'
          type: array
          title: Items
        total:
          type: integer
          title: Total
        offset:
          type: integer
          title: Offset
        limit:
          type: integer
          title: Limit
      type: object
      required:
      - items
      - total
      - offset
      - limit
      title: AppointmentListResponse
      description: Paginated list response model for appointments.
...
    AppointmentResponse:
      properties:
        patient_id:
          type: string
          format: uuid
          title: Patient Id
        therapist_id:
          type: string
          format: uuid
          title: Therapist Id
        scheduled_at:
          type: string
          format: date-time
          title: Scheduled At
        duration_minutes:
          type: integer
          exclusiveMinimum: 0.0
          title: Duration Minutes
          default: 30
        status:
          allOf:
          - $ref: '#/components/schemas/AppointmentStatus'
          default: scheduled
        source:
          allOf:
          - $ref: '#/components/schemas/AppointmentSource'
          default: manual
        id:
          type: string
          format: uuid
          title: Id
        clinic_id:
          type: string
          format: uuid
          title: Clinic Id
        created_at:
          type: string
          format: date-time
          title: Created At
        updated_at:
          type: string
          format: date-time
          title: Updated At
      type: object
      required:
      - patient_id
      - therapist_id
      - scheduled_at
      - id
      - clinic_id
      - created_at
      - updated_at
      title: AppointmentResponse
      description: Response model for an appointment.
    AppointmentSource:
      type: string
      enum:
      - manual
      - public_booking
      title: AppointmentSource
      description: Sources from which appointments are created.
    AppointmentStatus:
      type: string
      enum:
      - scheduled
      - completed
      - cancelled
      - no_show
      title: AppointmentStatus
      description: Scheduling states for appointments.
    AppointmentUpdate:
      properties:
        patient_id:
          anyOf:
          - type: string
            format: uuid
          - type: 'null'
          title: Patient Id
        therapist_id:
          anyOf:
          - type: string
            format: uuid
          - type: 'null'
          title: Therapist Id
        scheduled_at:
          anyOf:
          - type: string
            format: date-time
          - type: 'null'
          title: Scheduled At
        duration_minutes:
          anyOf:
          - type: integer
            exclusiveMinimum: 0.0
          - type: 'null'
          title: Duration Minutes
        status:
          anyOf:
          - $ref: '#/components/schemas/AppointmentStatus'
          - type: 'null'
        source:
          anyOf:
          - $ref: '#/components/schemas/AppointmentSource'
          - type: 'null'
      type: object
      title: AppointmentUpdate
      description: Payload for updating an existing appointment.
...
    HTTPValidationError:
      properties:
        detail:
          items:
            $ref: '#/components/schemas/ValidationError'
          type: array
          title: Detail
      type: object
      title: HTTPValidationError
...
    ValidationError:
      properties:
        loc:
          items:
            anyOf:
            - type: string
            - type: integer
          type: array
          title: Location
        msg:
          type: string
          title: Message
        type:
          type: string
          title: Error Type
      type: object
      required:
      - loc
      - msg
      - type
      title: ValidationError
  securitySchemes:
    OAuth2PasswordBearer:
      type: oauth2
      flows:
        password:
          scopes: {}
          tokenUrl: /api/v1/auth/login